Move from server to lib. Server was unecesarily
