--- servers/sql_server/sql_server.cpp
@@ -1,6 +1,0 @@
-#include "sql_server.h"
-
-int sql::server_run(int argc, char* argv[]){
-    std::cerr << "Starting SQL Server" << std::endl;
-    return 0;
-}
