+++ servers/sqllib/CMakeLists.txt
@@ -1,15 +1,19 @@
+cmake_minimum_required(VERSION 3.10)
+
+project(sqllib)
+
 # add the library that runs
-set(sql_server_source_files
-	sql_server.cpp
+set(sqllib_source_files
+	sqllib.cpp
 )
-add_library(sql_server ${sql_server_source_files})
+add_library(sqllib ${sqllib_source_files})
 
 # state that anybody linking to us needs to include the current source dir
-# to find MathFunctsqlns.h, while we don't.
-target_include_directories(sql_server
+target_include_directories(sqllib
                            INTERFACE
                             $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
                             $<INSTALL_INTERFACE:include>
                            )
 
-target_link_libraries(sql_server PUBLIC)
+target_link_libraries(sqllib PUBLIC)
+
