+++ servers/sqllib/sqllib.cpp
@@ -1,0 +1,6 @@
+#include "sqllib.h"
+
+int sqllib::test(){
+    std::cout << "SQLLib" << std::endl;
+    return 0;
+}
